Fix a number of bugs and add new user view.
Make admin a permitted parameter for creating and updating users (but only
admins may set it to true if it was not already). Also fix errors related to
creating new computers, which previously didn't allow a non-admin to create
computers. Now a user may make computers for themselves, but not others.
Admins may create computers for anyone. Finally, the user show action now
shows a nicer view of details, computers, and recently-submitted test
instances.